moving link condition back to trimeshop to take advantage or protectedness
