+++ src/wmtk/TriMeshOperation.cpp
@@ -373,160 +373,6 @@ std::string TriMeshSmoothVertexOperation::name() const
 // }
 
 
-auto TriMeshEdgeCollapseOperation::execute(TriMesh& m, const Tuple& loc0) -> ExecuteReturnData
-{
-    ExecuteReturnData ret_data;
-    std::vector<Tuple>& new_tris = ret_data.new_tris;
-    Tuple& return_t = ret_data.tuple;
-
-    auto& vertex_connectivity = this->vertex_connectivity(m);
-    auto& tri_connectivity = this->tri_connectivity(m);
-    // get fid for the return tuple
-    // take the face that shares the same vertex the loc0 tuple is pointing to
-    // or if that face doesn't exit
-    // take the face that shares the same vertex of loc0
-    size_t new_fid;
-    {
-        std::optional<Tuple> new_tup_opt;
-        new_tup_opt = loc0.switch_vertex(m).switch_edge(m).switch_face(m);
-        if (!new_tup_opt.has_value()) {
-            new_tup_opt = loc0.switch_edge(m).switch_face(m);
-            assert(new_tup_opt.has_value());
-        }
-        new_fid = new_tup_opt.value().fid(m);
-    }
-
-    // get the vids
-    size_t vid1 = loc0.vid(m);
-    size_t vid2 = m.switch_vertex(loc0).vid(m);
-
-
-    // get the fids
-    auto n1_fids = vertex_connectivity[vid1].m_conn_tris;
-
-    auto n2_fids = vertex_connectivity[vid2].m_conn_tris;
-
-    // get the fids that will be modified
-    auto n12_intersect_fids = set_intersection(n1_fids, n2_fids);
-    // check if the triangles intersection is the one adjcent to the edge
-    size_t test_fid1 = loc0.fid(m);
-    TriMesh::Tuple loc1 = m.switch_face(loc0).value_or(loc0);
-    size_t test_fid2 = loc1.fid(m);
-    //"faces at the edge is not correct"
-    assert(
-        vector_contains(n12_intersect_fids, test_fid1) &&
-        vector_contains(n12_intersect_fids, test_fid2));
-    // now mark the vertices as removed so the assertion for tuple validity in switch operations
-    // won't fail
-    vertex_connectivity[vid1].m_is_removed = true;
-    vertex_connectivity[vid2].m_is_removed = true;
-    for (size_t fid : n12_intersect_fids) {
-        tri_connectivity[fid].m_is_removed = true;
-    }
-
-    std::vector<size_t> n12_union_fids;
-    std::set_union(
-        n1_fids.begin(),
-        n1_fids.end(),
-        n2_fids.begin(),
-        n2_fids.end(),
-        std::back_inserter(n12_union_fids));
-
-    // record the fids that will be modified/erased for roll back on failure
-    vector_unique(n12_union_fids);
-    std::vector<std::pair<size_t, TriangleConnectivity>> old_tris(n12_union_fids.size());
-
-    for (const size_t fid : n12_union_fids) {
-        tri_connectivity[fid].hash++;
-    }
-    // modify the triangles
-    // the m_conn_tris needs to be sorted
-    size_t new_vid = get_next_empty_slot_v(m);
-    for (size_t fid : n1_fids) {
-        if (tri_connectivity[fid].m_is_removed)
-            continue;
-        else {
-            int j = tri_connectivity[fid].find(vid1);
-            tri_connectivity[fid].m_indices[j] = new_vid;
-        }
-    }
-    for (size_t fid : n2_fids) {
-        if (tri_connectivity[fid].m_is_removed)
-            continue;
-        else {
-            int j = tri_connectivity[fid].find(vid2);
-            tri_connectivity[fid].m_indices[j] = new_vid;
-        }
-    }
-
-    // now work on vids
-    // add in the new vertex
-
-    for (size_t fid : n12_union_fids) {
-        if (tri_connectivity[fid].m_is_removed)
-            continue;
-        else
-            vertex_connectivity[new_vid].m_conn_tris.push_back(fid);
-    }
-    vertex_connectivity[new_vid].m_is_removed = false;
-    // This is sorting too, and it is important to sort
-    vector_unique(vertex_connectivity[new_vid].m_conn_tris);
-
-    // remove the erased fids from the vertices' (the one of the triangles that is not the end
-    // points of the edge) connectivity list
-    std::vector<std::pair<size_t, size_t>> same_edge_vid_fid;
-    for (size_t fid : n12_intersect_fids) {
-        auto f_vids = tri_connectivity[fid].m_indices;
-        for (size_t f_vid : f_vids) {
-            if (f_vid != vid1 && f_vid != vid2) {
-                same_edge_vid_fid.emplace_back(f_vid, fid);
-                assert(vector_contains(vertex_connectivity[f_vid].m_conn_tris, fid));
-                vector_erase(vertex_connectivity[f_vid].m_conn_tris, fid);
-            }
-        }
-    }
-
-    // ? ? tuples changes. this needs to be done before post check since checked are done on tuples
-    // update the old tuple version number
-    // create an edge tuple for each changed edge
-    // call back check will be done on this vector of tuples
-
-    assert(vertex_connectivity[new_vid].m_conn_tris.size() != 0);
-
-    const size_t gfid = vertex_connectivity[new_vid].m_conn_tris[0];
-    int j = tri_connectivity[gfid].find(new_vid);
-    auto new_t = Tuple(new_vid, (j + 2) % 3, gfid, m);
-    int j_ret = tri_connectivity[new_fid].find(new_vid);
-    return_t = Tuple(new_vid, (j_ret + 2) % 3, new_fid, m);
-    assert(new_t.is_valid(m));
-
-    new_tris = m.get_one_ring_tris_for_vertex(new_t);
-
-    ret_data.success = true;
-    return ret_data;
-}
-
-bool TriMeshEdgeCollapseOperation::check_link_condition(const TriMesh& mesh, const Tuple& edge)
-{
-    return utils::link_condition(mesh, edge);
-}
-
-
-bool TriMeshEdgeCollapseOperation::before(TriMesh& m, const Tuple& t)
-{
-    return check_link_condition(m, t);
-}
-
-bool TriMeshEdgeCollapseOperation::after(TriMesh& m, ExecuteReturnData& ret_data)
-{
-    ret_data.success &= true;
-    return ret_data;
-}
-
-std::string TriMeshEdgeCollapseOperation::name() const
-{
-    return "edge_collapse";
-}
 
 auto TriMeshConsolidateOperation::execute(TriMesh& m, const Tuple& t) -> ExecuteReturnData
 {
